refactor: SignUpScreen에 modifier 파라미터 삭제
