+++ app/src/main/java/nextstep/signup/ui/SignUpActivity.kt
@@ -4,18 +4,13 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import kotlinx.coroutines.launch
 import nextstep.signup.ui.theme.SignupTheme
 
@@ -30,13 +25,8 @@ class SignUpActivity : ComponentActivity() {
                 Scaffold(
                     modifier = Modifier.fillMaxSize(),
                     snackbarHost = { SnackbarHost(snackbarHostState) },
-                ) { paddingValues ->
+                ) {
                     SignUpScreen(
-                        modifier =
-                            Modifier
-                                .fillMaxSize()
-                                .padding(paddingValues)
-                                .verticalScroll(rememberScrollState()),
                         snackBarEvent = { message ->
                             coroutineScope.launch {
                                 snackbarHostState.showSnackbar(
@@ -50,10 +40,4 @@ class SignUpActivity : ComponentActivity() {
             }
         }
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SignUpPreview() {
-    SignUpScreen()
 }
