+++ app/src/main/java/nextstep/signup/ui/SignUpScreen.kt
@@ -2,9 +2,12 @@ package nextstep.signup.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -15,6 +18,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import nextstep.signup.R
 import nextstep.signup.ui.component.SignUpHeaderText
@@ -27,12 +31,12 @@ import nextstep.signup.ui.model.SignUpInformation
 import nextstep.signup.ui.model.UserName
 
 @Composable
-fun SignUpScreen(
-    modifier: Modifier = Modifier,
-    snackBarEvent: (String) -> Unit = {},
-) {
+fun SignUpScreen(snackBarEvent: (String) -> Unit = {}) {
     Column(
-        modifier = modifier,
+        modifier =
+            Modifier
+                .fillMaxSize()
+                .verticalScroll(rememberScrollState()),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         var userName by rememberSaveable { mutableStateOf(UserName()) }
@@ -146,4 +150,10 @@ fun SignUpScreen(
 
         Spacer(modifier = Modifier.padding(bottom = 28.dp))
     }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun SignUpPreview() {
+    SignUpScreen()
 }
